Prevent nullpointer dereference in `Touch::getPosition`
If no finger is down, the `find` call on a `std::unordered_map` will
return `unordered_map::end` which should not be dereferenced. Thus
resulting in a crash.

This commit adds a check before dereferencing the result of `find` and
returns an empty `Vector2i` if it is `unordered_map::end`.

Issue: https://github.com/SFML/SFML/issues/2257